Fixed the following warning from GCC 4.0: mess/vidhrdw/i82720.c: In function 'compis_gdc_start': mess/vidhrdw/i82720.c:1414: warning: case label value exceeds maximum value for type (0x20000 is too large for a UINT16 type.)

+++ mess/vidhrdw/i82720.h
@@ -45,7 +45,7 @@ typedef enum
 typedef struct compis_gdc_interface
 {
 	compis_gdc_modes mode;
-	UINT16	vramsize;
+	UINT32	vramsize;
 } compis_gdc_interface;
 
 extern void mdrv_compisgdc(struct InternalMachineDriver *machine,
